\name{tuhocr-package}
\alias{tuhocr-package}
\alias{tuhocr}
\docType{package}
\title{
Functions for VN users
}
\description{
www.tuhocr.com
}
\details{
\tabular{ll}{
Package: \tab tuhocr\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2023-02-25\cr
License: \tab Free for all\cr
}
}
\author{
Duc Nguyen | Founder of tuhocr.com

Chuyên đào tạo kỹ năng xử lý dữ liệu sử dụng phần mềm R

Maintainer: Duc Nguyen <tuhocr.com@gmail.com>
}


