\name{tuhocr-package}
\alias{tuhocr-package}
\alias{tuhocr}
\docType{package}
\title{
Functions for daily tasks
}
\description{
Package này tập hợp các function giúp giải quyết công việc xử lý dữ liệu thường ngày. Để biết cách viết function, thân mời bạn tham gia khóa học R ở www.tuhocr.com giúp trang bị kiến thức R vững chắc.
}
\details{
\tabular{ll}{
Package: \tab tuhocr\cr
Type: \tab Package\cr
Version: \tab 0.1.2\cr
Date: \tab 2023-06-22\cr
License: \tab GPL (>= 3)\cr
}
}
\author{
Duc Nguyen | Founder of tuhocr.com

Chuyên đào tạo kỹ năng xử lý dữ liệu sử dụng phần mềm R cho người mới bắt đầu.

Maintainer: Duc Nguyen <tuhocr.com@gmail.com>
}


